task filter modal rework
