+++ forms-flow-review/src/actions/actionConstants.ts
@@ -15,7 +15,7 @@ export default {
     DEFAULT_FILTER: "DEFAULT_FILTER",
     BPM_FILTER_LIST: "BPM_FILTER_LIST",
     IS_BPM_FILTERS_LOADING: "IS_BPM_FILTERS_LOADING",
-    BPM_SELECTED_FILTER: "BPM_SELECTED_FILTER",
+    SET_SELECTED_FILTER: "SET_SELECTED_FILTER",
     SELECTED_TASK_ID: "SELECTED_TASK_ID",
     UPDATE_FILTER_SEARCH_PARAMS:"UPDATE_FILTER_SEARCH_PARAMS",
     TASKLIST: "TASKLIST",
