+++ forms-flow-review/src/types/task_filter_variables.ts
@@ -1,0 +1,10 @@
+export interface TaskVariables{
+    key: string;
+    width?: number | string;
+    label: string;
+    type?: string;
+    name: string;
+    isChecked: boolean;
+    sortOrder: number;
+    isFormVariable: boolean;
+}
